Added local cart to database cart upon login

+++ TCGshopTestEnvironment/Controllers/ProductsController.cs
@@ -98,7 +98,7 @@ namespace TCGshopTestEnvironment.Controllers
                 {
                     Id = result.prods.ProductId,
                     Name = result.prods.Name,/*.Length < 20 ? result.prods.Name : result.prods.Name.Substring(0, 15) + "...",*/
-                    Price = (decimal)result.prods.Price,
+                    Price = result.prods.Price,
                     ImageUrl = result.prods.ImageUrl,
                     Grade = result.prods.Grade,
                     Stock = result.prods.Stock,
@@ -233,7 +233,7 @@ namespace TCGshopTestEnvironment.Controllers
                     {
                         Id = result.prods.ProductId,
                         Name = result.prods.Name,/*.Length < 20 ? result.prods.Name : result.prods.Name.Substring(0, 15) + "...",*/
-                        Price = (decimal)result.prods.Price,
+                        Price = result.prods.Price,
                         ImageUrl = result.prods.ImageUrl,
                         Grade = result.prods.Grade,
                         Stock = result.prods.Stock,
